refactor: extract SettingForm
