+++ src/components/pages/select/SettingModal.tsx
@@ -10,17 +10,14 @@ import {
   ModalHeader,
   ModalOverlay,
   Text,
-  VStack,
 } from "@chakra-ui/react";
 import { Dispatch } from "react";
-import { CheckboxOptions } from "@/components/CheckboxOptions";
-import { SettingNumberOption } from "./SettingNumberOption";
-import { SettingSwitchOption } from "./SettingSwitchOption";
-import { BALLS_CONTENT, BallType } from "@/constants/balls";
+import { BALLS_CONTENT } from "@/constants/balls";
 import { InfoOutlineIcon } from "@chakra-ui/icons";
 import { useRouter } from "next/router";
 import { OptionsType } from "@/types";
 import { useGlobalState } from "@/hooks/useGlobalState";
+import { SettingForm } from "./SettingForm";
 
 interface OwnProps {
   open: boolean;
@@ -36,23 +33,7 @@ export const SettingModal = ({
   dispatch,
 }: OwnProps) => {
   const router = useRouter();
-  const toggleSwitch = (
-    type: Exclude<
-      SelectAction["type"],
-      "versions" | "numberOfQuiz" | "selectCard" | "reset"
-    >
-  ) => {
-    dispatch({ type, value: !options[type] });
-  };
   const { globalStateDispatch } = useGlobalState();
-
-  const updateNumberOfQuiz = (num: number) => {
-    dispatch({ type: "numberOfQuiz", value: num });
-  };
-
-  const updateVersion = (option: OptionsType["versions"]) => {
-    dispatch({ type: "versions", value: option });
-  };
 
   const onClickSubmit = () => {
     globalStateDispatch({ type: "updateOptions", value: options });
@@ -89,72 +70,9 @@ export const SettingModal = ({
               <Text fontSize={"sm"}>{ballContent.description}</Text>
             </Flex>
           )}
-          <Text fontSize={"lg"} fontWeight={700} my={2}>
-            出題形式
-          </Text>
-          <VStack spacing={4} align={"left"} my={2}>
-            <SettingNumberOption
-              title="出題数"
-              value={options["numberOfQuiz"]}
-              onChange={updateNumberOfQuiz}
-              disabled={!isCustom}
-            />
-            <SettingSwitchOption
-              title="4択で出題する"
-              value={options["isChoice"]}
-              onChange={() => toggleSwitch("isChoice")}
-              disabled={!isCustom}
-            />
-            <SettingSwitchOption
-              title="ヒントを表示する"
-              value={options["showHint"] && !options["isChoice"]}
-              onChange={() => toggleSwitch("showHint")}
-              disabled={!isCustom || options["isChoice"]}
-            />
-            <SettingSwitchOption
-              title="シルエットで出題する"
-              value={options["isSilhouette"]}
-              onChange={() => toggleSwitch("isSilhouette")}
-              disabled={!isCustom}
-            />
-          </VStack>
-          <Text fontSize={"lg"} fontWeight={700} my={2}>
-            出題範囲
-          </Text>
-          <VStack spacing={4} align={"left"} my={2}>
-            <SettingSwitchOption
-              title="リージョンフォームを含む"
-              caption="例：ニャース（ガラルのすがた）など"
-              value={options["hasRegion"]}
-              onChange={() => toggleSwitch("hasRegion")}
-              disabled={!isCustom}
-            />
-            <SettingSwitchOption
-              title="フォルム違いを含む"
-              caption="例：デオキシス（アタックフォルム）など"
-              value={options["hasAnotherForm"]}
-              onChange={() => toggleSwitch("hasAnotherForm")}
-              disabled={!isCustom}
-            />
-            <SettingSwitchOption
-              title="メガシンカを含む"
-              value={options["hasMega"]}
-              onChange={() => toggleSwitch("hasMega")}
-              disabled={!isCustom}
-            />
-            <SettingSwitchOption
-              title="キョダイマックスを含む"
-              value={options["hasGigantic"]}
-              onChange={() => toggleSwitch("hasGigantic")}
-              disabled={!isCustom}
-            />
-            <CheckboxOptions
-              title="初登場シリーズで絞り込む"
-              options={options.versions}
-              updateOptions={updateVersion}
-              disabled={!isCustom}
-            />
-          </VStack>
+          {/* TODO: 難易度の説明を実装する */}
+          {/* {ballContent ?? <BallSetting></BallSetting>} */}
+          {isCustom && <SettingForm dispatch={dispatch} options={options} />}
         </ModalBody>
         <ModalFooter columnGap={4}>
           <Button colorScheme={"orange"} onClick={onClickSubmit}>
